new tests

+++ src/map/mod.rs
@@ -78,7 +78,7 @@ async fn add_map(
 
     let z = Marker {
         id: None,
-        created_by: user.id.unwrap() as i64,
+        created_by: user.id.unwrap(),
         solved_by: None,
         point: PointC {
             v: Point {
@@ -126,7 +126,7 @@ sql_function!(fn resolve_marker(marker_id: BigInt, user_id: BigInt));
 #[post("/resolve/<marker_id>")]
 async fn resolve_marker_from_id(marker_id: i64, user: User, connection: Db) -> Status {
     let y = connection
-        .run(move |conn| select(resolve_marker(marker_id, user.id.unwrap() as i64)).execute(conn))
+        .run(move |conn| select(resolve_marker(marker_id, user.id.unwrap())).execute(conn))
         .await;
     if let Err(tmp) = y {
         match tmp.to_string().as_str() {
